Fixed bug where subcommands with no key arguments would still ask for one
Moved the subparser creation to CLITool so now all subparsers store their name in the args namespace as 'subparser_name'. This lets CLITool validate key arguments depending on the subcommand, which means key_parameters is now a dictionary instead of a list. Each key of the dictionary is a subcommand, and each value is a list of key parameters